HHH-18284 Add pre/post-upsert listener classes to StaticClassLists.typesNeedingArrayCopy
