+++ hibernate-graalvm/src/main/java/org/hibernate/graalvm/internal/StaticClassLists.java
@@ -84,10 +84,12 @@ final class StaticClassLists {
 				org.hibernate.event.spi.PreLoadEventListener[].class,
 				org.hibernate.event.spi.PreDeleteEventListener[].class,
 				org.hibernate.event.spi.PreUpdateEventListener[].class,
+				org.hibernate.event.spi.PreUpsertEventListener[].class,
 				org.hibernate.event.spi.PreInsertEventListener[].class,
 				org.hibernate.event.spi.PostLoadEventListener[].class,
 				org.hibernate.event.spi.PostDeleteEventListener[].class,
 				org.hibernate.event.spi.PostUpdateEventListener[].class,
+				org.hibernate.event.spi.PostUpsertEventListener[].class,
 				org.hibernate.event.spi.PostInsertEventListener[].class,
 				org.hibernate.event.spi.PreCollectionRecreateEventListener[].class,
 				org.hibernate.event.spi.PreCollectionRemoveEventListener[].class,
